pyproject.toml
[project]
name = "cloudy-poro-backend"
version = "2.0.0"
description = "Cloudy Poro - Professional esports scouting and analysis API for League of Legends teams"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "scikit-learn>=1.4.0",
    "requests>=2.32.5",
    "websockets>=12.0",
]

[project.scripts]
cloudy-poro = "src.main:app"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
